requirements.txt
tensorflow==2.0.3
keras==2.3.1
pillow
numpy
protobuf==3.20.0
matplotlib
pandas